rename, fix some lint stuff
