+++ frontend/src/components/transactions/EmissionsFactorFilter.tsx
@@ -5,7 +5,7 @@ import EmissionsFactorSelector from "./CategorySelector";
 import { useEffect, useState } from "react";
 import { cn } from "@/lib/utils";
 import { useAuth } from "@/context/AuthContext";
-import { useTransactionsContext } from "@/context/TableContext";
+import { useTransactionsContext } from "@/context/TransactionsContext";
 
 export default function EmissionsFactorFilter({
   className,
