--- frontend/src/context/LineItemsContext.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { GetLineItemResponse, LineItemFilters } from "@/types";
-import { createDataContext } from "./Context";
-
-export const { DataProvider: LineItemsProvider, useData: useLineItems } =
-  createDataContext<GetLineItemResponse, LineItemFilters>();
